Add init function for store, more consistency for getting commands and forgot ScriptEditor
