+++ packages/server-web/src/client/create-app.js
@@ -2,12 +2,12 @@ import Vue from "vue";
 import VueResource from "vue-resource";
 import AppComponent from "./component/App.vue";
 import { initRouter } from "./create-router.js";
-import { store } from "./store.js";
+import { initStore } from "./create-store";
 import "../style/scaffolding.scss";
 
 export default function createApp(commands) {
   Vue.use(VueResource);
-  store.commit("commands", commands);
+  const store = initStore({ Vue, commands });
   const router = initRouter({ Vue });
   new Vue({
     el: "#app",
